requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
decorator==4.4.2
Flask==3.0.2
gunicorn==21.2.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.5
moviepy==1.0.3
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==24.0
pillow==10.2.0
proglog==0.1.10
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
requests==2.31.0
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.1
